Wrote '03_prepare_filesystems.sh" script to fomrmat, partition and mount the drive.
The scripts in xtools were taken from: 'https://github.com/leahneukirchen/xtools/tree/master'
and are normally provided in the void linux live iso. They help with chroot and writing the
mount points to /etc/fstab. There is no reason to re-implement these and
as they are packaged with xbps, we won't be able to install them on a
non-void linux system. Since we are using the 'tarball' method, im
simply including them in my repo to make my life easier.